Move code only called on Linux into a Linux-specific file
Only moves unchanged code, should not change behavior.
